+++ mrv2/lib/mrvApp/mrvMainControl.cpp
@@ -189,7 +189,7 @@ namespace mrv
 
         p.displayOptions = value;
 
-        _widgetUpdate();
+        _displayUpdate();
     }
 
     void MainControl::setImageOptions(const timeline::ImageOptions& value)
@@ -214,6 +214,27 @@ namespace mrv
         }
     }
 
+    void MainControl::_displayUpdate()
+    {
+        TLRENDER_P();
+        
+        Viewport* view = p.ui->uiView;
+        view->setLUTOptions(p.lutOptions);
+        view->setDisplayOptions({p.displayOptions});
+        if (p.ui->uiSecondary)
+        {
+            view = p.ui->uiSecondary->viewport();
+            view->setLUTOptions(p.lutOptions);
+            view->setDisplayOptions({p.displayOptions});
+        }
+        
+        if (panel::colorPanel)
+        {
+            panel::colorPanel->setLUTOptions(p.lutOptions);
+            panel::colorPanel->setDisplayOptions(p.displayOptions);
+        }
+    }
+    
     void MainControl::_compareUpdate()
     {
         TLRENDER_P();