+++ mrv2/lib/mrvApp/mrvMainControl.h
@@ -29,6 +29,7 @@ namespace mrv
         void _timelinePlayersUpdate();
         void _widgetUpdate();
         void _compareUpdate();
+        void _displayUpdate();
 
         TLRENDER_PRIVATE();
     };